feat: duration of the track will be shown even in offline
I have found a way how we display the duration of the current track in the UI, even if we are offline and the track could not load.

Till now we were displaying the duration after loading the track, playing it and setting it to the currentTrack in reducer. And we used to display it through the TrackPlayer's useProgress() hook. which will provide the duration only when the track is available.

Now what I have tweaked is, when any track need to play. I am updating its state even before playing it. So that the duration of the track gets updated on the reducers storage. And I am displaying the duration from the state itself (the currentTrack.duration)