music util update below:
this is a simple update but fixes many issues
the artist data was throwing the error below:

undefined is not a object x.hasOwnProperty is undefined

and this lead to not rendering and returning the artist data. and data where some thumbnail issues are there. to fix this the changes are done.