screen: the actual searching screen
Ok, so there are few problems. When a search is submitted we have to replace the whole search history component, genre list component, and search suggestion's component with the search results.

Which is a hactic and perfomative task, So why not when the submit of search is performed, We will send the user to a new screen where the search results will be displayed.

In this way we won't be replacing the whole screen's UI. Instead we will preserve it and show the results separately.

There is one more solution to this, That we make all the search results as absolute. So that the results are displayed on top of the actual search screen UI. But this method will make the component so huge and more complex. So the first solution is best.