very major: albums are deprecated from sobyte
may be albums will be available in future, but for now they are deprecated and we are not providing albums data to the user

why this now, its becuase in most albums there is only one song. So why to show that album, we will show the song instead. Right!