major feat: now song can be played from any screen
HURRAY! This is a milestone of this application.
We are able to play any track from anywhere in the application with no hustle anymore.

Algorithm:
while playing any song we are providing extra description like the song data, screen data, and the query (by which the song is produced). After getting this we are setting this to the description of the track player before playing the song data. This description is provided in the metadata of the track player in TrackPlayer.add() function in string format.

Now an event is created which when the track is changed gets triggered. We are parsing the string description to JSON. Then we are checking the screen data from the metadata and getting if the song is not played from the player interface itself. Then we are executing the query in the description. And filtering out the currenly added song by the user. And adding all the rest of songs to the queue.