major: the track's url are saved in local storage
Ok previously after getting the url from origin, we were saving it to the reducer only. And when we open the app it loads the data again from the origin and then from reducer. But it seems that every URL is valid for maximum of 6 hours. But just to be in safe side. I made the URLs so that they will be saved for 5 hours in the local storage, with expiration data, url and musicID.

When any link request is made we will first check if that is in the reducer and then check in the local storage. If it is there provide the URL and update in reducer. If not then make a API request and save it to reducer after providing the URL.

NOTE: the expiration time is important to provide when saving URL to the local storage. Since when we are saving the data we are adding 5 hours to the current time. And while getting that data we are subtracting the current time from expiration time (previously generated time + 5 hours). And if it is less than 0 than we know that the URL is broken.

Also NOTE: I have taken 5 hours so that it is always less than 6 which is the maximum expiration period of any track's URL.

So, In other words - Offline cache capability is made possible. Cache is handled by TrackPlayer and URL is by Local storage.